Dev Answer: Prevent Refresh & pass props
This commit answers the two previous developer questions. A prop is
passed down to the the TaskForm component which allows it to insert a
task by calling the list modifier function insertTask().

The other issue is fixed by handling the submit in TaskForm correctly
and calling e.preventDefault() on the onSubmit event which automaticaly
refreshes the page and must be disabled to maintain data in the client